Limit CPU percent panels in cloud view to 100%
Set the y-max limit to 100 on panels that show a percentage. In some
cases a cloud restart or re-connection to STF after a period of no data
can result in a spike of the graph, resulting in a momentary peak of
several million percent.

This change sets the graphs to always show a range of 0 to 100% to limit
the visual impact of an overscale / high peak value problem.

Related: rhbz#2006970