Enhance SG tracker dashboard
Enhance the SG tracker dashboard by filling the background colour of the
thresholds defined for message receive count. If the message count
largely exceeds the threshold, the visual layer can become compressed in
such a way that it's not clear the threshold is being overrun.
Additionally, add a new panel row that shows the collectors and
transports memory usage.