Set VM dashboard annontation view off by default
Change the annotations for virtual machine checks in the virtual machine
dashboard view off by default. The Elasticsearch instance is not enabled
by default in our documentation or in a default ServiceTelemetry
manifest, so having this enabled by default can result in errors showing
up in the dashboard when you don't have the es_ceilometer data source.

Also changed the overlay colour from green to grey because the green
seemed a bit intense when looking at a 24 hour range (which is the
dashboard default).

Closes: rhbz#2173856